pyproject.toml
[tool.poetry]
name = "lipstick"
version = "0.1.0"
description = ""
authors = ["Stephen Nicholas Swatman"]
readme = "README.md"
packages = [{include = "lipstick"}]

[tool.poetry.dependencies]
python = "^3.11"
z3-solver = "^4.12.4.0"
networkx = "^3.2.1"
pydantic = "^2.5.3"
pyyaml = "^6.0.1"
rich = "^13.7.0"
matplotlib = "^3.8.2"

[tool.poetry.scripts]
lipstick = "lipstick.cli:main"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ipython = "^8.12.0"
flake8 = "^6.0.0"
flake8-black = "^0.3.6"
isort = "^5.12.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
